+++ ExcelTableConverter.LatexTableConverter.Tests/LatexConverterTests.cs
@@ -44,7 +44,7 @@ namespace ExcelTableConverter.LatexTableConverter.Tests
 
       var latexConverter = new LatexConverter(extendedFeaturesModel);
 
-      var expected = "\\begin{table}[!ht]\r\n\\centering\r\n\\begin{tabular}{lllll}\r\n\r\n1 & 20% & 1 3 & 1 4 & 1 5 \\\\\r\n2 1 & 2 2 & 2 3 & 2 4 & 2 5 \\\\\r\n3 1 & 3 2 & 3 3 & 3 4 & 3 5 \\\\\r\n4 1 & 4 2 & 4 3 & 4 4 & 4 5 \\\\\r\n\r\n\\end{tabular}\r\n\\caption{SheetName}\r\n\\end{table}";
+      var expected = "\\begin{table}\r\n\\centering\r\n\\begin{tabular}{lllll}\r\n\r\n1 & 20% & 1 3 & 1 4 & 1 5 \\\\\r\n2 1 & 2 2 & 2 3 & 2 4 & 2 5 \\\\\r\n3 1 & 3 2 & 3 3 & 3 4 & 3 5 \\\\\r\n4 1 & 4 2 & 4 3 & 4 4 & 4 5 \\\\\r\n\r\n\\end{tabular}\r\n\\caption{SheetName}\r\n\\end{table}";
       var result = latexConverter.GetConvertedContent(Constants.GetTableWithAllFeatures());
       Assert.That(result, Is.EqualTo(expected));
     }
@@ -133,7 +133,7 @@ namespace ExcelTableConverter.LatexTableConverter.Tests
 
       var latexConverter = new LatexConverter(extendedFeaturesModel);
 
-      var expected = "\\begin{table}[!ht]\r\n\\renewcommand{\\arraystretch}{1.2}\r\n\\centering\r\n\\begin{tabular}{@{}lllll@{}}\r\n\\toprule\r\n1 & 20% & 1 3 & 1 4 & 1 5 \\\\\\midrule\r\n2 1 & 2 2 & 2 3 & 2 4 & 2 5 \\\\\r\n3 1 & 3 2 & 3 3 & 3 4 & 3 5 \\\\\r\n4 1 & 4 2 & 4 3 & 4 4 & 4 5 \\\\\\bottomrule\r\n\r\n\\end{tabular}\r\n\\caption{SheetName}\r\n\\end{table}";
+      var expected = "\\begin{table}\r\n\\renewcommand{\\arraystretch}{1.2}\r\n\\centering\r\n\\begin{tabular}{@{}lllll@{}}\r\n\\toprule\r\n1 & 20% & 1 3 & 1 4 & 1 5 \\\\\\midrule\r\n2 1 & 2 2 & 2 3 & 2 4 & 2 5 \\\\\r\n3 1 & 3 2 & 3 3 & 3 4 & 3 5 \\\\\r\n4 1 & 4 2 & 4 3 & 4 4 & 4 5 \\\\\\bottomrule\r\n\r\n\\end{tabular}\r\n\\caption{SheetName}\r\n\\end{table}";
       var result = latexConverter.GetConvertedContent(Constants.GetTableWithAllFeatures());
       Assert.That(result, Is.EqualTo(expected));
     }